+++ CHANGELOG.md
@@ -1,13 +1,19 @@
 # Changelog
 
-## [1.0.0@next] - 2023-04-19
+## [1.0.0@next] - 2023-04-21
+- Fix OpenAI analysis generation when using proxy backend.
+- Add support to export and import reports.
+- Add many missing french translations.
 - Fix OpenAI analysis generation when using proxy backend.
 - Add Symbols/Random to open the pattern view of a random stock (just for fun!)
-- First release
 
 # TODO
 
-- [ ] 2023-04-19: Add Export/Import of reports
+- [ ] 2023-04-20: The search window should also use the /api/search endpoint to search for symbols combine to the search index
+- [ ] 2023-04-20: The search window should allow to add a symbol to a report
+- [ ] 2023-04-20: Use the new search window to add a new title
+- [ ] 2023-04-20: Add a + button on the report table background to add a new title
+- [ ] 2023-04-20: Check if we can download a first search index database from the backend
 - [ ] 2023-04-19: Add user guide
 - [ ] 2023-04-19: Add user authentication when using the proxy backend
 - [ ] 2023-04-19: Add augmented logo cache from the backend